postgres_to_es/requirements.txt
python-dotenv~=0.21.0
psycopg2-binary~=2.9
pytz~=2022.6
sqlparse~=0.4.3
elasticsearch~=8.5.3
pydantic~=1.10.2